Fix leaked loggers
In the NodeLauncher we currently create a specific logger per node we
create. However loggers cannot be destroyed and will live for the
whole lifetime of the application. This means that we essentially leak
one Logger object per node we create.

The reason for introducing per node loggers was that we can easily
filter the log by node id and get every message related to that node
creation. This also can be accomplished by using a LoggerAdapter which
is also already in use by zuul for the same purpose.

Change-Id: Ieff99c242c1ec1ca89e2c16929cfc3237d6f1a78