Explicitly set use_direct_get to False
It is essential for nodepool that we use list and local filtering
for openstack servers for scale reasons. For users who are using
openstacksdk at scales that are not nodepool, it's less efficient
to do lists all the time. There is a flag to control the behavior
of get_server. Set it explicitly to request the behavior nodepool
is looking for so that openstacksdk can change the default to
better serve non-nodepool consumers.

Change-Id: I7cf03285f9a04f3eef403c67d75e149605207eb1