Allow nodepool for using in-cluster configs
When running within a kubernetes cluster, it's easier to configure a
service account and mounting it on the nodepool POD than creating a
kubeconfig file and making it available to nodepool.

For this reason, this commit adds the ability to load configs from the
in-cluster service account paths. It does this as a fallback when the
kubeconfig path doesn't exist.

This commit also makes `context` a non required configuration option,
since it's not needed when a service account is used.

Change-Id: I7762940993c185c17d7468df72dff22e99d7f8c2